+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # x.y.z
 
-- Add prop `confirmComponent` for custom confirm dialog
+- Add customization props `confirmComponent`, `previewField` & `reviewField`
 - Remove unnecessary aria label from print button
 
 # 0.4.93